chores: fix 32bit build with -Wconversion

+++ libsqsh/src/archive/trailing_context.c
@@ -43,7 +43,7 @@ sqsh__trailing_init(
 	const struct SqshSuperblock *superblock = sqsh_archive_superblock(sqsh);
 	uint64_t trailing_start = sqsh_superblock_bytes_used(superblock);
 	struct SqshMapManager *map_manager = sqsh_archive_map_manager(sqsh);
-	size_t archive_size = sqsh__map_manager_size(map_manager);
+	uint64_t archive_size = sqsh__map_manager_size(map_manager);
 	uint64_t trailing_size;
 
 	if (archive_size <= trailing_start) {
